changed cookiereader class to be more generic and follow current style
